Update to Kotlin 2.4.20, JUnit 5.14.4, kotest 5.9.1, mockk 1.14.11, Dokka 2.2.0 (#162)
* Update Kotlin to 2.4.20, JUnit to 5.14.4, kotest to 5.9.1, mockk to 1.14.11, Dokka to 2.2.0.

Use the junit-jupiter aggregate artifact and declare the JUnit platform
launcher explicitly, and disable kotest's classpath autoscan.

* Remove redundant toLong() calls flagged by Kotlin 2.4.

+++ src/main/kotlin/org/jitsi/utils/Instant.kt
@@ -40,11 +40,11 @@ import java.time.Instant
  */
 fun Instant.toEpochMicro(): Long {
     if (epochSecond < 0 && nano > 0) {
-        val millis = Math.multiplyExact(epochSecond + 1, 1_000_000).toLong()
+        val millis = Math.multiplyExact(epochSecond + 1, 1_000_000)
         val adjustment: Long = (nano / 1_000 - 1).toLong()
         return Math.addExact(millis, adjustment)
     } else {
-        val millis = Math.multiplyExact(epochSecond, 1_000_000).toLong()
+        val millis = Math.multiplyExact(epochSecond, 1_000_000)
         return Math.addExact(millis, (nano / 1000).toLong())
     }
 }
